requirements.txt
praw
tweepy
requests
python-dotenv